requirements.txt
azure-functions
azure-storage-blob>=12.0.0
pytest
